feat: Added JWT Token authentication to the code
