+++ accounts/token.py
@@ -1,6 +1,10 @@
+import datetime
+from functools import wraps
+from accounts.models import APIKey, User
 from django.contrib.auth.tokens import PasswordResetTokenGenerator
 from django.utils import six
-
+from django.conf import settings as st
+from jwt import encode, decode
 
 class TokenGenerator(PasswordResetTokenGenerator):
     def _make_hash_value(self, user, timestamp):
@@ -11,3 +15,13 @@ class TokenGenerator(PasswordResetTokenGenerator):
 
 
 account_activation_token = TokenGenerator()
+
+jwt_secret = st.JWT_SECRET
+
+
+def get_jwt_token(user: User) -> str:
+    return encode({'id': user.user_id, 'email': user.email, "exp": datetime.datetime.utcnow() + datetime.timedelta(hours=24)}, jwt_secret, algorithm="HS256")
+
+def decode_jwt_token(token: str) -> User or None:
+    return decode(token, jwt_secret)
+
