+++ gasdeliverybackend/auth_token.py
@@ -4,3 +4,6 @@ AUTH_STEP = 8
 SECRET_KEY = 'xfgcotz(dm+vco3qozc9bd+5%(%1)k*x%_@wnyuiqhk7&9-cnq'
 PAYSTACK_SECRET = 'sk_test_7d09abaa42aa10d3c129a376252da7672146f119'
 PAYSTACK_PUBLIC = 'pk_test_84c0906d6adef0ad69d00c10830304848688a797'
+
+TZN = 'Africa/Lagos'
+AUTH_PROTOCOL = 'jwt'
